Maddr.email converted to BinaryField.
